Move Create New Post components to its own component, and also structured the files
